client/requirements.txt
grpcio-tools
simple-hexdump
